pyproject.toml
[project]
name = "blog-api"
version = "0.1.0"
description = ""
authors = [
    {name = "adrlksv",email = "Qwerty123Andrey@bk.ru"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[all] (>=0.115.8,<0.116.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "bcrypt (==4.0.1)",
    "alembic (>=1.14.1,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
]

[tool.poetry]
packages = [{include = "blog_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
